Attempted to add daemonization, but it is trickier then expected.  Appears close to working.

+++ ynot/pid.py
@@ -11,10 +11,13 @@ def cleanupPid(fileName=defaultPidFile):
 
 def makePidFile(fileName=defaultPidFile):
     pid = os.getpid()
-    f = open(fileName, "w")
-    contents = "%s\n"%(pid,)
-    f.write(contents)
-    f.close()
-    atexit.register(cleanupPid, fileName)
+    if not os.path.exists(fileName):
+        f = open(fileName, "w")
+        contents = "%s\n"%(pid,)
+        f.write(contents)
+        f.close()
+        atexit.register(cleanupPid, fileName)
+    else:
+        raise RuntimeError("%s pid file already exists, this file is either stale or there is already an instance of ynot running"%(fileName,))
